pyproject.toml
[tool.poetry]
name = "vara-backend"
version = "0.1.0"
description = ""
authors = ["Niwaiwa <po2846@gmail.com>"]
readme = "README.md"
packages = [{include = "vara_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "~4.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
